requirements.txt

langchain
openai
chromadb
tiktoken
python-dotenv
